+++ src/commands/createFolderStructure.ts
@@ -11,9 +11,14 @@ import {
   readConfig,
 } from "../lib/vscodeHelpers";
 import { showError, showInfo, getWorkspaceUri } from "../lib/vscodeHelpers";
-import { getTemplatesFromFS, pickTemplate, replaceTemplateContent } from "../lib/extensionHelpers";
+import {
+  getTemplatesFromFS,
+  pickTemplate,
+  replaceTemplateContent,
+} from "../lib/extensionHelpers";
 import { fileExists, getFullFilePath, isDirectory } from "../lib/fsHelpers";
 import { relative } from "path";
+import { chmodSync } from "fs";
 
 const CreateFolderStructure = async (
   resource: vscode.Uri | string | undefined,
@@ -22,7 +27,10 @@ const CreateFolderStructure = async (
   const workspaceUri = await getWorkspaceUri();
   const targetUri = await getTargetPath(resource, workspaceUri);
 
-  const templateFolderPaths = [await getLocalTemplatePath(targetUri), await getGlobalTemplatePath(globalTemplatePath)];
+  const templateFolderPaths = [
+    await getLocalTemplatePath(targetUri),
+    await getGlobalTemplatePath(globalTemplatePath),
+  ];
 
   const validPaths = templateFolderPaths.filter(isDirectory) as string[];
 
@@ -50,11 +58,13 @@ const CreateFolderStructure = async (
     openFilesWhenDone,
     omitParentDirectory = false,
     omitFTName = false,
-    overwriteExistingFiles = 'never',
+    overwriteExistingFiles = "never",
     absolutePath = false,
+    setExecutablePermission = false,
   } = pickedTemplate;
 
-  const templateNotation = pickedTemplate.templateNotation || readConfig("templateNotation");
+  const templateNotation =
+    pickedTemplate.templateNotation || readConfig("templateNotation");
 
   if (omitFTName && !omitParentDirectory) {
     return showError(
@@ -70,52 +80,114 @@ const CreateFolderStructure = async (
 
   let ftNameTuple: StringReplaceTuple[] = [];
   if (!omitFTName) {
-    try{
+    try {
       ftNameTuple = await getReplaceValueTuples(["FTName"]);
       //If no componentname is specified do nothing
       if (!ftNameTuple[0][1]) {
         return showInfo(
           "Aborted folder creation. Cannot continue without a name"
-          );
-        }
-    } catch(e: any){
-      if(e && e.message === 'Cancelled'){
+        );
+      }
+    } catch (e: any) {
+      if (e && e.message === "Cancelled") {
         return showInfo("Aborted folder creation. Cancelled input");
       }
     }
   }
 
   //Get all inputs for replacement of customvariables
-  const replaceValueTuples = ftNameTuple.concat(await getReplaceValueTuples(customVariables || []));
+  const replaceValueTuples = ftNameTuple.concat(
+    await getReplaceValueTuples(customVariables || [])
+  );
   const fsPath = (absolutePath ? workspaceUri : targetUri)?.fsPath;
-  const structureContents = files.map(row => ({
-    fileName: getFullFilePath(row.fileName, fsPath, replaceValueTuples, omitParentDirectory, templateNotation),
-    template: replaceTemplateContent(row.template, replaceValueTuples, templateNotation)
+  const structureContents = files.map((row) => ({
+    isExecutable: row.isExecutable,
+    fileName: getFullFilePath(
+      row.fileName,
+      fsPath,
+      replaceValueTuples,
+      omitParentDirectory,
+      templateNotation
+    ),
+    template: replaceTemplateContent(
+      row.template,
+      replaceValueTuples,
+      templateNotation
+    ),
   }));
 
   let filesToCreate: FolderStructure;
 
-  if (overwriteExistingFiles === 'always') {
+  if (overwriteExistingFiles === "always") {
     filesToCreate = structureContents;
-  } else if(overwriteExistingFiles === 'prompt'){
+  } else if (overwriteExistingFiles === "prompt") {
     //Doesn't work yet. Fix when coming back. Files to create are figured out in a wrong way
     const existingFiles = structureContents.filter(fileExists);
-    const newFiles = structureContents.filter(file => !fileExists(file));
-    const filesToOverwrite = existingFiles.length ? await vscode.window.showQuickPick(existingFiles.map(row => ({...row, label: relative(fsPath || '', row.fileName)})), {canPickMany:true, placeHolder: 'Please select the files that should be overwritten'}) || [] : [];
-    filesToCreate = newFiles.concat(structureContents.filter(file => filesToOverwrite.find(row => row.fileName === file.fileName)));
+    const newFiles = structureContents.filter((file) => !fileExists(file));
+    const filesToOverwrite = existingFiles.length
+      ? (await vscode.window.showQuickPick(
+          existingFiles.map((row) => ({
+            ...row,
+            label: relative(fsPath || "", row.fileName),
+          })),
+          {
+            canPickMany: true,
+            placeHolder: "Please select the files that should be overwritten",
+          }
+        )) || []
+      : [];
+    filesToCreate = newFiles.concat(
+      structureContents.filter((file) =>
+        filesToOverwrite.find((row) => row.fileName === file.fileName)
+      )
+    );
   } else {
-    filesToCreate = structureContents.filter(file => !fileExists(file));
-    structureContents.filter(fileExists).forEach(file => showInfo(`Skipped creating file ${file.fileName} because it exists already`));
+    filesToCreate = structureContents.filter((file) => !fileExists(file));
+    structureContents
+      .filter(fileExists)
+      .forEach((file) =>
+        showInfo(
+          `Skipped creating file ${file.fileName} because it exists already`
+        )
+      );
   }
 
-  await createStructure(
-    filesToCreate,
-    templateNotation
-  );
+  const createdFiles = await createStructure(filesToCreate, templateNotation);
+
+  if (setExecutablePermission && createdFiles && createdFiles.length > 0) {
+    createdFiles.forEach(
+      (
+        file:
+          | {
+              filePath: string;
+              isExecutable: boolean;
+            }
+          | {
+              isExecutable: boolean;
+              filePath?: undefined;
+            }
+          | null
+      ) => {
+        if (file && file.isExecutable && file.filePath) {
+          chmodSync(file.filePath, "755");
+        }
+      }
+    );
+  }
 
   if (openFilesWhenDone && fsPath) {
     await Promise.all(
-      openFilesWhenDone.map(file => openFile(getFullFilePath(file, fsPath, replaceValueTuples, omitParentDirectory, templateNotation)))
+      openFilesWhenDone.map((file) =>
+        openFile(
+          getFullFilePath(
+            file,
+            fsPath,
+            replaceValueTuples,
+            omitParentDirectory,
+            templateNotation
+          )
+        )
+      )
     );
   }